streamlit-ui/requirements.txt
streamlit>=1.28.0
requests>=2.28.0
boto3>=1.28.0
pyyaml>=6.0
